apps/python-api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5

# OpenTelemetry - Métricas
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-exporter-prometheus==0.50b0
prometheus-client==0.21.1

# OpenTelemetry - Tracing
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation-httpx==0.50b0
httpx==0.28.1

# Database (PostgreSQL)
psycopg2-binary==2.9.10
sqlalchemy==2.0.36

# Logging Estruturado
python-json-logger==2.0.7
